app/requirements.txt
asgiref==3.9.1
black==25.1.0
click==8.2.1
colorama==0.4.6
Django==5.2.4
djangorestframework==3.16.0
drf-excel==2.5.3
et_xmlfile==2.0.0
flake8==7.3.0
flake8-docstrings==1.7.0
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
platformdirs==4.3.8
psycopg2==2.9.10
pycodestyle==2.14.0
pydocstyle==6.3.0
pyflakes==3.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
snowballstemmer==3.0.1
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
